less flicker on initial play through slideshow
Switch lazy-loading strategy from 'ondemand' to 'progressive'.
This means that the image gallery now asyncrhonously loads all images
after the component is initially rendered.

A downside of this strategy is that if you skip to the middle of the
slideshow you may have to wait for those images to be downloaded, as they
are no longer loaded "on demand".