Stream new timesteps as they arrive
Poll the Girder API, checking for new timesteps. When they appear, add
them to our image gallery. Automatically advance to the latest timestep
if the user was watching the end of the steram.

As of this commit we no longer loop back to the beginning of the
slide show when we reach the last timestep. Instead we stay at the end
waiting for more data to arrive.