Return to previous location on selection clear
If a suggested parameter has been selected the location is updated to display
that parameter. When the clear icon is selected the location will be updated to
the previous location before the selection was made. This does not change
location if the clear icon is used to erase the query data entered by the user,
only when a selection is made from the drop-down list of matches.

Signed-off-by: Brianna Major <brianna.major@kitware.com>