Poll for new metadata on initial load
The max time steps were intially being set by the number of images found for the
first selected parameter rather than the metadata. This would set an inaccurate
max value if that parameter had less images than the actual number of steps. For
the first parameter selected poll the metadata for the actual max.

Signed-off-by: Brianna Major <brianna.major@kitware.com>