Do not assume images are sequential when loading
Instead of finding the image in the array by index that is determined by the time
step, find the correct image by looking for the object whose step key matches the
current time step. If there is no image for the current time step select the
previous image and display that until there is an image that matches the time
step.

Signed-off-by: Brianna Major <brianna.major@kitware.com>