Add option to set range for json data
For json data the option to set the global range is now added to the context
menu when you right click on the parameter's plot. The range cannot be set for
static images. Clearing and saving the range will reset the plots and
automatically scale the y axis based on the current timestep data.

Signed-off-by: Brianna Major <brianna.major@kitware.com>