Support loading a View as a template
If an existing View was created from data in the same simulation but different
run than the current run, an option to load the View as a template will be
enabled. If the current run does not contain one or more of the parameters
included in the template View the cell will be left empty. The timestep
associated with the View is ignored and is instead reset to step 1.

Signed-off-by: Brianna Major <brianna.major@kitware.com>