Re-design the notifications display
This now more closely resembles to download notification dialog that Google
Drive uses. There is a single snackbar item that is displayed with a list of
reuested downloads that reflect the current status (in progress, completed,
failed). They can be manually dismissed or will be automatically cleared 5
seconds after completion or failure.

Signed-off-by: Brianna Major <brianna.major@kitware.com>