Add additional "download options" option to context menu
The conext menu now contains a "download options" option which opens a dialog.
The user can use the dialog to select whether they want to download a movie or
images, the format they'd like and whether they would like to download all
time steps or a list and/or range of specific time steps.

Signed-off-by: Brianna Major <brianna.major@kitware.com>