Improve state management by managing each plot state separately
Originally we were managing states with lots of objects, each managing
references to the currently loaded plots and the state of that plot. Now we
have a module instance that is created each time a plot is added to the
dashboard. This instance manages all of the plot state information and can be
accessed with the the unique item id fetched from Girder.

Signed-off-by: Brianna Major <brianna.major@kitware.com>