Cancel pending tasks for plot when a new template is loaded
Previously pending tasks would sometimes resolve after a plot was replaced by a
template and plot data for the previous selection would be loaded. This was
causing the wrong plot to be displayed or no plot to render at all when the plot
type had changed. Make sure we cancel pending tasks now if the item id has
changed.

Signed-off-by: Brianna Major <brianna.major@kitware.com>